--- walletlibrary/src/main/java/com/microsoft/walletlibrary/did/sdk/identifier/IdentifierManager.kt
@@ -1,15 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-*  Copyright (c) Microsoft Corporation. All rights reserved.
-*  Licensed under the MIT License. See License.txt in the project root for license information.
-*--------------------------------------------------------------------------------------------*//*
-
-
-package com.microsoft.walletlibrary.did.sdk.identifier
-
-import com.microsoft.walletlibrary.did.sdk.identifier.models.Identifier
-import com.microsoft.walletlibrary.did.sdk.util.controlflow.Result
-
-internal interface IdentifierManager {
-    suspend fun getMasterIdentifier(): Result<Identifier>
-}
-*/
