--- walletlibrary/src/main/java/com/microsoft/walletlibrary/identifier/IdentifierManager.kt
@@ -1,15 +1,0 @@
-/*
-package com.microsoft.walletlibrary.identifier
-
-import com.microsoft.walletlibrary.did.sdk.IdentifierService
-import com.microsoft.walletlibrary.did.sdk.identifier.IdentifierManager
-import com.microsoft.walletlibrary.did.sdk.identifier.models.Identifier
-import com.microsoft.walletlibrary.did.sdk.util.controlflow.Result
-
-internal class IdentifierManager constructor(
-    private val identifierService: IdentifierService
-): IdentifierManager {
-    override suspend fun getMasterIdentifier(): Result<Identifier> {
-        return identifierService.getMasterIdentifier()
-    }
-}*/
