--- walletlibrary/src/test/java/com/microsoft/walletlibrary/did/sdk/RevocationServiceTest.kt
@@ -1,71 +1,0 @@
-/*
-// Copyright (c) Microsoft Corporation. All rights reserved
-
-package com.microsoft.walletlibrary.did.sdk
-
-import com.microsoft.walletlibrary.did.sdk.credential.models.RevocationReceipt
-import com.microsoft.walletlibrary.did.sdk.credential.models.VerifiableCredential
-import com.microsoft.walletlibrary.did.sdk.credential.service.models.RevocationRequest
-import com.microsoft.walletlibrary.did.sdk.credential.service.protectors.RevocationResponseFormatter
-import com.microsoft.walletlibrary.did.sdk.datasource.network.credentialOperations.SendVerifiablePresentationRevocationRequestNetworkOperation
-import com.microsoft.walletlibrary.did.sdk.identifier.models.Identifier
-import com.microsoft.walletlibrary.did.sdk.util.controlflow.Result
-import kotlin.Result as KotlinResult
-import io.mockk.coEvery
-import io.mockk.coVerify
-import io.mockk.mockk
-import io.mockk.mockkConstructor
-import io.mockk.spyk
-import kotlinx.coroutines.runBlocking
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.Before
-import org.junit.Test
-
-class RevocationServiceTest {
-
-    private val identifierService: IdentifierService = mockk()
-    private val masterIdentifier: Identifier = mockk()
-    private val revocationResponseFormatter: RevocationResponseFormatter = mockk()
-    private val revocationService =
-        spyk(RevocationService(mockk(relaxed = true), identifierService, revocationResponseFormatter, mockk()), recordPrivateCalls = true)
-
-    private val revokeRpList = listOf("did:ion:test")
-    private val revokeReason = "test reason"
-    private val verifiableCredential: VerifiableCredential = mockk()
-    private val formattedResponse = "FORMATTED_RESPONSE"
-
-    @Before
-    fun setup() {
-        mockkConstructor(SendVerifiablePresentationRevocationRequestNetworkOperation::class)
-        coEvery { identifierService.getMasterIdentifier() } returns Result.Success(masterIdentifier)
-        coEvery { verifiableCredential.contents.vc.revokeService?.id } returns "https://microsoft.com/vcs"
-        coEvery { revocationResponseFormatter.formatResponse(any(), any()) } returns formattedResponse
-    }
-
-    @Test
-    fun `revoke verifiable presentation successfully with correct params`() {
-        val expectedRevocationRequest = RevocationRequest(verifiableCredential, masterIdentifier, revokeRpList, revokeReason)
-        val expectedRevocationReceipt: RevocationReceipt = mockk()
-        coEvery { revocationService["sendRevocationRequest"](expectedRevocationRequest, formattedResponse) } returns Result.Success(
-            expectedRevocationReceipt
-        )
-
-        runBlocking {
-            val actualResult = revocationService.revokeVerifiablePresentation(verifiableCredential, revokeRpList, revokeReason)
-            assertThat(actualResult).isInstanceOf(Result.Success::class.java)
-            assertThat((actualResult as Result.Success).payload).isEqualTo(expectedRevocationReceipt)
-        }
-
-        coVerify(exactly = 1) {
-            revocationService["sendRevocationRequest"](expectedRevocationRequest, formattedResponse)
-        }
-    }
-
-    @Test
-    fun `passing empty list results in failure`() {
-        runBlocking {
-            val actualResult = revocationService.revokeVerifiablePresentation(verifiableCredential, emptyList(), revokeReason)
-            assertThat(actualResult).isInstanceOf(Result.Failure::class.java)
-        }
-    }
-}*/
